chore: move getV2ClientNames to helpers
